Move file getMonthlyMaxEstimated.
